refactor: move IAttributeParsedResult and derived classes to AttributeParsedResults namespace
Moved IAttributeParsedResult, its extensions, and derived classes (ParsedBitFiledAttribute, ParsedExtendedBitFiledAttribute, ParsedReadOnlyBitFieldAttribute, and related interfaces) to the PropertyBitPack.SourceGen.Models.AttributeParsedResults namespace. Updated using directives across the project to reflect the new structure.
