+++ PropertyBitPack.SourceGen/BitFieldPropertyAggregators/BaseUnnamedFieldAggregator.cs
@@ -2,6 +2,7 @@
 using Microsoft.CodeAnalysis;
 using PropertyBitPack.SourceGen.Collections;
 using PropertyBitPack.SourceGen.Models;
+using PropertyBitPack.SourceGen.Models.AttributeParsedResults;
 using PropertyBitPack.SourceGen.Models.FieldRequests;
 using PropertyBitPack.SourceGen.Models.GenerateSourceRequest;
 using System;
