+++ PropertyBitPack.SourceGen/Models/AttributeParsedResults/IParsedExtendedBitFiledAttribute.cs
@@ -1,6 +1,6 @@
 ﻿using Microsoft.CodeAnalysis;
 
-namespace PropertyBitPack.SourceGen.Models;
+namespace PropertyBitPack.SourceGen.Models.AttributeParsedResults;
 
 /// <summary>
 /// Represents the parsed data of an extended bit field attribute.
