+++ PropertyBitPack.SourceGen/PropertyBitPackGeneratorContext.cs
@@ -13,6 +13,7 @@ using System.Diagnostics.CodeAnalysis;
 using System.Text;
 using Microsoft.CodeAnalysis.Text;
 using PropertyBitPack.SourceGen.Models.GenerateSourceRequest;
+using PropertyBitPack.SourceGen.Models.AttributeParsedResults;
 
 namespace PropertyBitPack.SourceGen;
 internal abstract partial class PropertyBitPackGeneratorContext
